click & try, translation text strings, orders payment method fix, tech support information, updated readme, organized code structure for gateways, fixed tags for proper html readability.
