--- plugins/modenapaymentgateway/assets/css/modena-admin-style.css
@@ -1,73 +1,0 @@
-fieldset label[for="woocommerce_modena_slice_enabled"],
-fieldset label[for="woocommerce_modena_direct_enabled"],
-fieldset label[for="woocommerce_modena_business_leasing_enabled"],
-fieldset label[for="woocommerce_modena_click_enabled"],
-fieldset label[for="woocommerce_modena_credit_enabled"],
-fieldset label[for="woocommerce_modena_slice_product_page_banner_enabled"],
-fieldset label[for="woocommerce_modena_direct_product_page_banner_enabled"],
-fieldset label[for="woocommerce_modena_business_leasing_product_page_banner_enabled"],
-fieldset label[for="woocommerce_modena_credit_product_page_banner_enabled"] {
-    position: relative;
-    display: inline-block;
-    width: 56px;
-    height: 30px;
-}
-
-fieldset label[for="woocommerce_modena_slice_enabled"] input,
-fieldset label[for="woocommerce_modena_direct_enabled"] input,
-fieldset label[for="woocommerce_modena_business_leasing_enabled"] input,
-fieldset label[for="woocommerce_modena_click_enabled"] input,
-fieldset label[for="woocommerce_modena_credit_enabled"] input,
-fieldset label[for="woocommerce_modena_slice_product_page_banner_enabled"] input,
-fieldset label[for="woocommerce_modena_business_leasing_product_page_banner_enabled"] input,
-fieldset label[for="woocommerce_modena_click_product_page_banner_enabled"] input,
-fieldset label[for="woocommerce_modena_direct_product_page_banner_enabled"] input,
-fieldset label[for="woocommerce_modena_credit_product_page_banner_enabled"] input {
-    opacity: 0;
-    width: 0;
-    height: 0;
-}
-
-.modena-slider {
-    position: absolute;
-    cursor: pointer;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: #ccc;
-    -webkit-transition: .4s;
-    transition: .4s;
-    border-radius: 30px;
-}
-
-.modena-slider:before {
-    position: absolute;
-    content: "";
-    height: 22px;
-    width: 22px;
-    left: 4px;
-    bottom: 4px;
-    background-color: white;
-    -webkit-transition: .4s;
-    transition: .4s;
-    border-radius: 50%;
-}
-
-input:checked + .modena-slider {
-    background-color: #2196F3;
-}
-
-input:focus + .modena-slider {
-    box-shadow: 0 0 1px #2196F3;
-}
-
-input:checked + .modena-slider:before {
-    -webkit-transform: translateX(26px);
-    -ms-transform: translateX(26px);
-    transform: translateX(26px);
-}
-
-.live_env_alert {
-    color: #de2a2a;
-}
