--- plugins/modenapaymentgateway/assets/css/modena-checkout.css
@@ -1,58 +1,0 @@
-.mdn_banklinks {
-    display: flex !important;
-    flex-direction: row !important;
-    align-content: center !important;
-    flex-wrap: wrap !important;
-    justify-content: center;
-}
-
-.mdn_banklink_img {
-    box-sizing: border-box !important;
-    min-width: 75px !important;
-    width: 75px !important;
-    max-width: 75px !important;
-    background: #ffffff !important;
-    padding: 0 !important;
-    margin: 5px !important;
-    cursor: pointer;
-    border-radius: 14px !important;
-    border: thin solid #eeeeee !important;
-    transition: all 150ms ease-out;
-}
-
-.mdn_banklink_img:hover {
-    background: #f2f2f2 !important;
-}
-
-.mdn_checked {
-    border: solid #20dfac 4px !important;
-}
-
-.mdn_service_info {
-    margin-left: 24px;
-    font-size: 12px;
-    color: #aaaaaa;
-    text-decoration: none;
-}
-
-.payment_method_modena_direct label {
-    display: inline-block;
-    vertical-align: middle;
-}
-
-.payment_method_modena_slice label,
-.payment_method_modena_business_leasing label,
-.payment_method_modena_click label,
-.payment_method_modena_credit label {
-    display: inline-block;
-    vertical-align: middle;
-    max-width: calc(100% - 33px);
-}
-
-.payment_method_modena_direct label img,
-.payment_method_modena_slice label img,
-.payment_method_modena_business_leasing label img,
-.payment_method_modena_click label img,
-.payment_method_modena_credit label img {
-    margin-left: 0 !important;
-}
