--- plugins/modenapaymentgateway/autoload.php
@@ -1,13 +1,0 @@
-<?php
-
-require_once __DIR__ . '/lib/Modena.php';
-require_once __DIR__ . '/lib/model/ModenaBaseModel.php';
-require_once __DIR__ . '/lib/model/ModenaApplication.php';
-require_once __DIR__ . '/lib/model/ModenaCustomer.php';
-require_once __DIR__ . '/lib/model/ModenaOrderItem.php';
-require_once __DIR__ . '/lib/model/ModenaRequest.php';
-require_once __DIR__ . '/lib/model/ModenaResponse.php';
-require_once __DIR__ . '/lib/model/ModenaOrderResponse.php';
-require_once __DIR__ . '/includes/class-modena-log-handler.php';
-require_once __DIR__ . '/includes/class-modena-onboarding-handler.php';
-require_once __DIR__ . '/includes/class-modena-load-checkout-assets.php';
