--- plugins/modenapaymentgateway/readme.txt
@@ -1,70 +1,0 @@
-=== Modena Payment Gateway ===
-Version: 2.7.0
-Contributors: modenapayments
-Donate link: https://modena.ee/
-Tags: online payment, payment, payment gateway, payment_gateway, woocommerce, sales, hire-purchase, järelmaks, financing, credit intermediary, krediidivahendaja, modena
-Requires at least: 4.7
-Tested up to: 6.1
-Stable tag: 2.7.0
-Requires PHP: 7.0
-License: GPLv2 or later
-License URI: https://www.gnu.org/licenses/gpl-2.0.html
-
-
-
-== Description ==
-
-Modena is a full checkout solution for all of your e-commerce needs. We cover all popular payment methods. Modena can help you get started with everything you need to start your online store checkout in Estonia.
-
-=== BANK LINKS AND CARD PAYMENTS===
-We help you cover all major banks in Estonia to collect your payments. All transactions are immediate. In addition we can provide your e-commerce site with card payments.
-
-=== FINANCING ===
-Modena offers a broad range of credit products to a wide variety of your target audience. We offer you an option accept business or personal credit offers for your product. We help you to make more business by making your products more available to a wider audience.
-
-=== SPLIT PAYMENTS - BUY NOW AND PAY LATER ===
-Your customers have the ability to pay for your services in 3 or less parts, making your services accessible to more clients.
-
-=== HOW TO GET STARTED ===
-1. Let us know if you want to sign up. Visit https://modena.ee.
-2. Verify your identity.
-3. Set up the plugin and start receiving payments.
-
-== Installation ==
-1. Create an account at [website](https://www.modena.ee) and install the plugin from Wordpress Plugin Repository [website](https://wordpress.org/plugins/modenapaymentgateway/)
-2. Find API Key and Secret at Modena Partner Portal [website](https://www.partner.modena.ee). Connect Modena Payment Option under Woocommerce Payments.
-3. Verify gateways and start accepting payments with Modena!
-
-If you have any questions about the payment solutions, then you can find additional integration information from our [developer website](https://developer.modena.ee/) or more information about Modena products can be found on our website [website](https://www.modena.ee).
-Any questions? Just send us an email at info@modena.ee.
-
-== Screenshots ==
-1. Bank Links & Card Payments
-2. Credit Payments
-3. Pay In 3
-
-== Changelog ==
-
-= 2.5.0 =
-* Initial wordpress plugin
-
-= 2.6.1 =
-* Bugfixes
-
-= 2.7.0 =
-* New Business Leasing payment option for merchants
-
-== Upgrade Notice ==
-
-= 2.5.0 =
-* Install initial version
-
-= 2.6.1 =
-* Bugfixes
-
-= 2.7.0 =
-* New Business Leasing payment option for merchants
-
-== Support ==
-More information about Modena products can be found on our website [website](https://www.modena.ee)
-Any questions? Just send us an email at info@modena.ee.
